# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "BISU-QA System API",
      "description": "棠心问答系统后端 API 文档",
      "version": "1.0.0",
      "contact": {
        "name": "BISU-QA 开发团队",
        "email": "support@example.com"
      }
    },
    "servers": [
      {
        "url": "http://10.10.15.210:5000",
        "description": "开发环境服务器"
      },
      {
        "url": "http://10.101.0.208:5000",
        "description": "测试环境服务器"
      }
    ],
    "paths": {
...
      "/api/auth/login": {
        "post": {
          "summary": "混合认证登录",
          "description": "支持RADIUS和本地数据库认证的登录接口",
          "operationId": "hybridLogin",
          "tags": ["用户认证"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["username", "password"],
                  "properties": {
                    "username": {
                      "type": "string",
                      "description": "用户名",
                      "example": "admin"
                    },
                    "password": {
                      "type": "string",
                      "description": "密码",
                      "example": "admin@123"
                    },
                    "auth_type": {
                      "type": "string",
                      "description": "认证类型 (radius/local/hybrid)",
                      "example": "local",
                      "enum": ["radius", "local", "hybrid"]
                    },
                    "nas_ip_address": {
                      "type": "string",
                      "description": "NAS-IP-Address (仅用于RADIUS认证)",
                      "example": "10.10.15.211"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "认证成功",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": true
                      },
                      "message": {
                        "type": "string",
                        "example": "登录成功"
                      },
                      "token": {
                        "type": "string",
                        "example": "dummy_token_admin"
                      },
                      "user": {
                        "type": "object",
                        "properties": {
                          "username": {
                            "type": "string",
                            "example": "admin"
                          },
                          "auth_method": {
                            "type": "string",
                            "example": "本地数据库"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "请求参数错误",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "401": {
              "description": "认证失败",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/auth/create-admin": {
        "post": {
          "summary": "创建管理员用户",
          "description": "创建新的管理员用户",
          "operationId": "createAdmin",
          "tags": ["用户认证"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["username", "password"],
                  "properties": {
                    "username": {
                      "type": "string",
                      "description": "用户名",
                      "example": "newadmin"
                    },
                    "password": {
                      "type": "string",
                      "description": "密码",
                      "example": "newpassword"
                    },
                    "email": {
                      "type": "string",
                      "description": "邮箱地址",
                      "example": "newadmin@example.com"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "创建成功",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": true
                      },
                      "message": {
                        "type": "string",
                        "example": "用户创建成功"
                      },
                      "user": {
                        "type": "object",
                        "properties": {
                          "username": {
                            "type": "string",
                            "example": "newadmin"
                          },
                          "email": {
                            "type": "string",
                            "example": "newadmin@example.com"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "请求参数错误",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "ErrorResponse": {
          "type": "object",
          "properties": {
            "status": {
              "type": "string",
              "example": "error"
            },
            "message": {
              "type": "string",
              "example": "发生错误"
            },
            "error_code": {
              "type": "string",
              "example": "SERVER_ERROR"
            }
          }
        },
...
      }
    },    "tags": [
...
  }